fixes/adjusted-file-and-reallocated-file
